back_end/requirements.txt
asgiref==3.5.2
bcrypt==4.0.0
boto3==1.24.76
botocore==1.27.76
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
decorator==5.1.1
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
dynaconf==3.1.9
hvac==1.0.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
paramiko==2.11.0
pycparser==2.21
pyhcl==0.4.4
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
validators==0.20.0
XlsxWriter==3.0.3
xlrd==1.2.0
django-jazzmin==2.5.0
